bleh.  test is passing, bug not found
added test for a zone with vds data, so the scale does kick in and it
will not sum to exactly number of days times daily aadt * miles of road
in highway.  However, the test does prove that the highway sections (in
the test case, highway 80 in Alameda County) in the HPMS data are
dropped out, and only the non-highway data are kept.

Still no bug found, all looks good.  Next try with more HPMS road
segments in the grid.