testing scratch work
figured out A bug, which might also be THE bug.  I was ignoring
direction, so if a road was bi directional in a grid cell, then it would
have twice the length, not just one times, so you could get twice the
total hpms.

I still recall the hpms sometimes being way higher than twice, but
anyway, that's a start.

new file too for testing purposes.