don't crash if a grid cell is empty
apparently, it may be possible to have an empty grid cell.  It gets
picked up by postgresql, but for whatever reason, it has no data in
couchdb to dictate the fractions to use (no modeling output).  In that
case, skip it, but don't die.

I haven't written the follow on test downstream yet