fixing things for one_hour case
various edits.

view emits key, null now, so use doc, not value

make sure nhh, not not_hh

data is an array of arrays, one array for each roadway with a detector.
The source data might be in error here, but regardless, the output of
this code will be array of arrays, as in

```
[
  [data, for, first, freeway]
  [data, for, second, freeway]
]
```