DO NOT multiply VMT by MT
I don't appear to be using this code, but to make sure everything is
consistent, delete the multiplication by detector_miles.

The couchdb database carb/grid/state4k was created by
calvad_precache_areas, which does the right thing and multiplies each
bit of volume for a roadway link by that link's length, as it goes.

See line 133 of lib/reducer.js in jmarca/calvad_reducer