year check.  didn't save it earlier
this function tacks on a year to the end of the couchdb database and
tests if the resulting database exists.  if it does, then it returns a
new config object with db holding the new, year-specific name.  If it
does not exist, then the original db name is returned.  If the incoming
db name has the same year as the testing year, then quit without hitting
couchdb.